refactor: changes after merge
