+++ web/pages/onboarding.tsx
@@ -1,0 +1,3 @@
+import { Onboarding } from "scenes/onboarding/onboarding";
+
+export default Onboarding;
